results/run-20230924-033218_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_5-runs_log/whatsapp/0/LLMatic/user.py
class User:
	def __init__(self, email, password):
		self.email = email
		self.password = password
		self.profile_picture = None
		self.status_message = None
		self.privacy_settings = None
		self.contacts = {}
		self.blocked_contacts = {}
		self.groups = {}
		self.status = None

	def sign_up(self, email, password):
		self.email = email
		self.password = password

	def recover_password(self, email):
		return self.password if self.email == email else None

	def set_profile_picture(self, picture):
		self.profile_picture = picture

	def set_status_message(self, message):
		self.status_message = message

	def set_privacy_settings(self, settings):
		self.privacy_settings = settings

	def block_contact(self, contact):
		self.blocked_contacts[contact] = self.contacts.pop(contact, None)

	def unblock_contact(self, contact):
		self.contacts[contact] = self.blocked_contacts.pop(contact, None)

	def create_group(self, group_name):
		self.groups[group_name] = []

	def edit_group(self, group_name, new_name):
		self.groups[new_name] = self.groups.pop(group_name, None)

	def manage_group(self, group_name, action, contact):
		if action == 'add':
			self.groups[group_name].append(contact)
		elif action == 'remove':
			self.groups[group_name].remove(contact)

	def send_message(self, recipient, message):
		return {'recipient': recipient, 'message': message}

	def receive_message(self, sender, message):
		return {'sender': sender, 'message': message}

	def read_message(self, message):
		return message

	def encrypt_message(self, message):
		return message

	def decrypt_message(self, message):
		return message

	def share_image(self, recipient, image):
		return {'recipient': recipient, 'image': image}

	def send_emoji(self, recipient, emoji):
		return {'recipient': recipient, 'emoji': emoji}

	def create_group_chat(self, group_name):
		self.groups[group_name] = []

	def add_participant(self, group_name, participant):
		self.groups[group_name].append(participant)

	def remove_participant(self, group_name, participant):
		self.groups[group_name].remove(participant)

	def assign_admin_role(self, group_name, participant):
		return {'group': group_name, 'admin': participant}

	def remove_admin_role(self, group_name, participant):
		return {'group': group_name, 'admin': participant}

	def post_status(self, status):
		self.status = status

	def view_status(self):
		return self.status

	def queue_message(self, recipient, message):
		return {'recipient': recipient, 'message': message}

	def display_online_status(self):
		return 'online'
